match: close OPEN floor-item renderer

+++ src/open/entity_render.c
@@ -113,7 +113,7 @@ void render_floor_item(KfFloorItem *item)
     SVECTOR screen;
     MATRIX model;
     long flag;
-    u8 facing;
+    u16 facing;
     u32 next_frame;
     u32 frame_count;
     s16 depth_bias;
@@ -125,7 +125,7 @@ void render_floor_item(KfFloorItem *item)
     screen.vz = (u16)item->position_z - (u16)open_graphics_runtime.render_state.view_position.vz;
     RotTrans(&screen, (VECTOR *)&model.t, &flag);
     facing = item->facing_and_frame_count & KF_FLOOR_ITEM_FACING_MASK;
-    if (facing != 0) {
+    if ((u8)facing != 0) {
         matrix_set_rotation_y(
             (facing - KF_FLOOR_ITEM_FACING_ZERO_YAW) << KF_FLOOR_ITEM_FACING_TO_ANGLE_SHIFT,
             &model);
